refactor: rename RGB types from rgb8_t/rgbf_t to rgba8_t/rgbaf_t
Updated all RGB type references throughout the codebase to include explicit alpha channel naming for consistency. This change affects:
- Core type definitions in rgb.hpp
- Conversion functions between color spaces
- Utility functions and operators
- Documentation and examples
- Test files and benchmarks

The new naming convention (rgba8_t, rgbaf_t) better reflects the actual 4-component structure of these types and maintains consistency with the existing rgba_t template.

+++ include/color/core/cmyk.hpp
@@ -85,7 +85,7 @@ inline constexpr cmyk_int_t cmyk_int = cmyk_int_t::make<C, M, Y, K>();
 /**
  * @brief Floating-point CMYK color type
  *
- * Represents CMYK colors using double precision floating-point values (0.0-1.0 range).
+ * Represents CMYK colors using float precision floating-point values (0.0-1.0 range).
  * Scale is set to 1000, mapping integer inputs to floating-point range.
  *
  * @tparam C Cyan component (scaled from integer to 0.0-1.0)
@@ -93,7 +93,7 @@ inline constexpr cmyk_int_t cmyk_int = cmyk_int_t::make<C, M, Y, K>();
  * @tparam Y Yellow component (scaled from integer to 0.0-1.0)
  * @tparam K Key (black) component (scaled from integer to 0.0-1.0)
  */
-using cmyk_float_t = basic_cmyk<double, 1000>;
+using cmyk_float_t = basic_cmyk<float, 1000>;
 template <int C, int M, int Y, int K>
 inline constexpr cmyk_float_t cmyk_float = cmyk_float_t::make<C, M, Y, K>();
